test: split generation handoff lifecycle coverage
